moved auth middleware to server.js Isn't able to verify the token...yet
