+++ api/middleware/auth.js
@@ -1,27 +1,29 @@
-require('dotenv').config();
+// require('dotenv').config();
 
-let admin = require('firebase-admin');
+// let admin = require('firebase-admin');
 
-let serviceAccount = require(process.env.GOOGLE_APPLICATION_CREDENTIALS);
+// let serviceAccount = require(process.env.GOOGLE_APPLICATION_CREDENTIALS);
 
-admin.initializeApp({
-    credential: admin.credential.cert(serviceAccount)
-    // databaseURL: 'https://voicechatroom-1874e.firebaseio.com'
-});
+// admin.initializeApp({
+//     credential: admin.credential.cert(serviceAccount)
+//     // databaseURL: 'https://voicechatroom-1874e.firebaseio.com'
+// });
 
-export const auth = (req, res, next) => {
-    var idToken = req.get('authorization');
+// const auth = (req, res, next) => {
+//     // var idToken = req.get('Authorization');
+//     var idToken = req.body;
 
-    // idToken comes from the client app
-    admin
-        .auth()
-        .verifyIdToken(idToken)
-        .then(function(decodedToken) {
-            let uid = decodedToken.uid;
-            res.locals.uid = uid;
-            next();
-        })
-        .catch(function(error) {
-            res.send(500).json({ message: 'Invalid token' });
-        });
-};
+//     // idToken comes from the client app
+//     admin
+//         .auth()
+//         .verifyIdToken(idToken)
+//         .then(function(decodedToken) {
+//             let uid = decodedToken.uid;
+//             res.locals.uid = uid;
+//             next();
+//         })
+//         .catch(function(error) {
+//             res.send(500).json({ message: 'Invalid token' });
+//         });
+// };
+// module.exports = auth;
