+++ package.json
@@ -21,7 +21,7 @@
     "ngrok": "^3.1.1",
     "node": "^11.12.0",
     "pg": "^7.9.0",
-    "sqlite3": "^4.0.6",
+    "sqlite3": "^4.0.9",
     "stripe": "6.25.1",
     "twilio": "^3.29.2"
   },
